Improved dialog layout.
There seems to be a bug in Qt4 Designer/uic. If you specify a layout margin of 0 in Designer, this will be converted to -1 by uic, which is the same as default, and not what you had in mind. Workaround: set a margin of 1 to get 0. Anyone else seen this?

svn path=/trunk/KDE/kdeedu/kwordquiz/; revision=471549