qml port: Port word quiz editor to QML
Currently missing the possibility to add sounds and images to a
KEduVocExpression.

A new feature introduced with this commit is that the document created
are stored in a list and stored in ~/.local/share/kwordquiz instead of
relying on KRecentFiles and letting people manage the files manually.

(cherry picked from commit 0b55f9113f4dacd30b6fac785e09cbe28007ad0d)